Make library flat again
