+++ src/windows_registry.rs
